+++ src/posts/entities/post.entity.ts
@@ -1,0 +1,43 @@
+import {
+  Column,
+  CreateDateColumn,
+  Entity,
+  ManyToOne,
+  OneToMany,
+  PrimaryGeneratedColumn,
+  UpdateDateColumn,
+} from 'typeorm';
+import { UserEntity } from '../../users/entities/user.entity';
+import { FileEntity } from '../../files/entities/file.entity';
+import { PostStatus } from '../post-status.enum';
+
+@Entity('posts')
+export class PostEntity {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column({ length: 255 })
+  title: string;
+
+  @Column('text')
+  content: string;
+
+  @ManyToOne(() => UserEntity, (user) => user.posts, { eager: true })
+  user: UserEntity;
+
+  @Column({
+    type: 'enum',
+    enum: PostStatus,
+    default: PostStatus.DRAFT,
+  })
+  status: PostStatus;
+
+  @OneToMany(() => FileEntity, (file) => file.post)
+  files: FileEntity[];
+
+  @CreateDateColumn()
+  createdAt: Date;
+
+  @UpdateDateColumn()
+  updatedAt: Date;
+}