+++ src/posts/posts.controller.ts
@@ -1,0 +1,63 @@
+import {
+  BadRequestException,
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Patch,
+  Post as HttpPost,
+  Request,
+  UseGuards,
+} from '@nestjs/common';
+import { PostsService } from './posts.service';
+import { CreatePostDto } from './dto/create-post.dto';
+import { UpdatePostDto } from './dto/update-post.dto';
+import { JwtAuthGuard } from '../auth/guard/jwt-auth.guard';
+import { PostVisibilityGuard } from './guards/post-visibility.guard';
+import { PostOwnerGuard } from './guards/post-owner.guard';
+import { RequestWithUser } from 'src/common/interfaces/request-with-user.interface';
+
+@Controller('posts')
+export class PostsController {
+  constructor(private readonly postsService: PostsService) {}
+
+  @HttpPost()
+  @UseGuards(JwtAuthGuard)
+  create(
+    @Body() createPostDto: CreatePostDto,
+    @Request() req: RequestWithUser,
+  ) {
+    return this.postsService.create(createPostDto, req.user);
+  }
+
+  @Get()
+  @UseGuards(JwtAuthGuard)
+  findAll() {
+    return this.postsService.findAll();
+  }
+
+  @Get(':id')
+  @UseGuards(JwtAuthGuard, PostVisibilityGuard)
+  findOne(@Param('id') id: string) {
+    return this.postsService.findOne(+id);
+  }
+
+  @Patch(':id')
+  @UseGuards(JwtAuthGuard, PostOwnerGuard)
+  update(@Param('id') id: string, @Body() updatePostDto: UpdatePostDto) {
+    return this.postsService.update(+id, updatePostDto);
+  }
+
+  @Delete(':id')
+  @UseGuards(JwtAuthGuard, PostOwnerGuard)
+  async remove(@Param('id') id: string) {
+    const resultDelete = await this.postsService.remove(+id);
+
+    if (!resultDelete) {
+      throw new BadRequestException('پست حذف نشد');
+    }
+
+    return { message: 'پست حذف شد' };
+  }
+}