+++ src/users/users.controller.ts
@@ -1,16 +1,17 @@
 import {
+  Body,
   Controller,
   Get,
   Patch,
-  Body,
-  UseGuards,
   Request,
+  UseGuards,
 } from '@nestjs/common';
 import { JwtAuthGuard } from '../auth/guard/jwt-auth.guard';
 import { UsersService } from './users.service';
 import { UpdateProfileDto } from './dto/update-profile.dto';
 import { UserProfileDto } from './dto/user-profile.dto';
 import { plainToInstance } from 'class-transformer';
+import { RequestWithUser } from '../common/interfaces/request-with-user.interface';
 
 @Controller('users')
 export class UsersController {
@@ -18,10 +19,8 @@ export class UsersController {
 
   @UseGuards(JwtAuthGuard)
   @Get('profile')
-  async getProfile(
-    @Request() req: { user: { userId: number } },
-  ): Promise<UserProfileDto> {
-    const userId = req.user.userId;
+  async getProfile(@Request() req: RequestWithUser): Promise<UserProfileDto> {
+    const userId = req.user.id;
 
     const user = await this.usersService.findById(userId);
 
@@ -31,10 +30,10 @@ export class UsersController {
   @UseGuards(JwtAuthGuard)
   @Patch('profile')
   async updateProfile(
-    @Request() req: { user: { userId: number } },
+    @Request() req: RequestWithUser,
     @Body() updateData: UpdateProfileDto,
   ): Promise<UserProfileDto> {
-    const userId: number = req.user.userId;
+    const userId: number = req.user.id;
 
     const updatedUser = await this.usersService.updateUser(userId, updateData);