+++ src/users/users.service.ts
@@ -1,26 +1,24 @@
-import {
-  Injectable,
-  BadRequestException,
-  ConflictException,
-} from '@nestjs/common';
+import { ConflictException, Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { Repository } from 'typeorm';
-import { User } from './entities/user.entity';
+import { UserEntity } from './entities/user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UpdateProfileDto } from './dto/update-profile.dto';
+import { FilesService } from '../files/files.service';
 
 @Injectable()
 export class UsersService {
   constructor(
-    @InjectRepository(User)
-    private readonly usersRepository: Repository<User>,
+    @InjectRepository(UserEntity)
+    private readonly usersRepository: Repository<UserEntity>,
+    private readonly filesService: FilesService,
   ) {}
 
-  create(createUserDto: CreateUserDto): Promise<User | undefined> {
+  create(createUserDto: CreateUserDto): Promise<UserEntity | undefined> {
     return this.usersRepository.save(createUserDto);
   }
 
-  findByEmail(email: string): Promise<User | null> {
+  findByEmail(email: string): Promise<UserEntity | null> {
     return this.usersRepository.findOne({ where: { email } });
   }
 
@@ -44,7 +42,7 @@ export class UsersService {
       return this.findById(id);
     }
 
-    // Check Field Uniques
+    // Check Field student_code Uniques
     if (updateData.student_code != undefined) {
       const existingUser = await this.usersRepository.findOne({
         where: { student_code: updateData.student_code },
@@ -53,6 +51,20 @@ export class UsersService {
       if (existingUser && existingUser.id !== id) {
         throw new ConflictException('کد دانشجویی تکراری است');
       }
+    }
+
+    // Check Field File Exist and Public
+    if (updateData.profile_file != undefined) {
+      const isFilePublic: boolean = await this.filesService.isPublic(
+        updateData.profile_file,
+      );
+
+      if (!isFilePublic) {
+        throw new ConflictException('فایل پروفایل شما غیر مجاز است');
+      }
+
+      // Mark File To Used
+      await this.filesService.markFileAsUsed(updateData.profile_file);
     }
 
     const cleanUpdateData: Record<string, any> = {};
@@ -67,7 +79,13 @@ export class UsersService {
       return this.findById(id);
     }
 
-    await this.usersRepository.update(id, cleanUpdateData);
+    await this.usersRepository.update(id, {
+      ...cleanUpdateData,
+      profile_file: cleanUpdateData.profile_file
+        ? { id: updateData.profile_file }
+        : null,
+    });
+
     return this.findById(id);
   }
 }