+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# [0.2.0]
+
+# fix
+
+- auth: fix jwt check guard
+
+# improvement
+
+- users: improve upload profile scenario
+- auth: encrypt password before save
+
+# feature
+
+- posts: add endpoints crud
+- files: add endpoints crud
+- common: add public interface and schedule clean files
+
 # [0.1.1]
 
 # feature